move packages to pkgs & kill eris-pm
